+++ src/obfs/http_simple.h
@@ -10,9 +10,9 @@
 struct obfs_t * http_simple_new_obfs();
 void http_simple_dispose(struct obfs_t *obfs);
 
-int http_simple_client_encode(struct obfs_t *obfs, char **pencryptdata, int datalength, size_t* capacity);
-int http_simple_client_decode(struct obfs_t *obfs, char **pencryptdata, int datalength, size_t* capacity, int *needsendback);
+size_t http_simple_client_encode(struct obfs_t *obfs, char **pencryptdata, size_t datalength, size_t* capacity);
+size_t http_simple_client_decode(struct obfs_t *obfs, char **pencryptdata, size_t datalength, size_t* capacity, int *needsendback);
 
-int http_post_client_encode(struct obfs_t *obfs, char **pencryptdata, int datalength, size_t* capacity);
+size_t http_post_client_encode(struct obfs_t *obfs, char **pencryptdata, size_t datalength, size_t* capacity);
 
 #endif // _OBFS_HTTP_SIMPLE_H
